requirements.txt
yfinance
pandas_ta
pandas
numpy
requests
python-dotenv
